Minor Update
Refined software performance. Software is now capable of operating using
only 30 megabytes of random access memories.
Also added a new window for the "Lost USB" feature. Note that this
feature is not functional.
The next release of the source code and executable will occur when the
"Lost USB" UI had been implemented.