Major Update ma01160907 patch 0.1a
This update introduces a new windows framework "Decrypting.vb", as well
as a new animation for the decryption. This is an experimental update,
and should only be used by developers as many users are experiencing
freezing issues on the decryption window (software will not proceed any
further).
This problem is still being researched. The first Beta release will be
released as soon as this freezing issue had been solved.
Patch 0.1a solves an issue where multiple pop ups will appear as soon as
user access the "Lost USB" page, accompanied with a system shutdown.